refactor: merge lib/shared into core, group sidebar/skills components
- Merged lib/ and shared/ into single core/ folder for simpler imports
- Moved slash-command.ts extension from components to core
- Grouped sidebar files into components/sidebar/ subfolder
- Grouped skill-related files into components/skills/ subfolder
- Updated all import paths to use @/core/ alias
- Fixed electron main/preload/skills-service imports
