+++ apps/desktop/src/components/file-manager-logo.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/core/utils";
 
 type FileManagerLogoProps = {
   className?: string;
