+++ apps/desktop/src/components/note-view.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useMemo } from "react";
 
-import { getDisplayFileName } from "@/lib/paths";
-import { getDirectTabShortcutDisplay, getShortcutDisplay } from "@/shared/shortcuts";
-import type { AppInfo, NoteTab, ShortcutSetting, TabMovePosition } from "@/shared/workspace";
+import { getDisplayFileName } from "@/core/paths";
+import { getDirectTabShortcutDisplay, getShortcutDisplay } from "@/core/shortcuts";
+import type { AppInfo, NoteTab, ShortcutSetting, TabMovePosition } from "@/core/workspace";
 import type { OutlineItem } from "@/types/navigation";
-import type { UpdateState } from "@/shared/workspace";
+import type { UpdateState } from "@/core/workspace";
 import type { EditorFindRequest, EditorFocusRequest } from "@/types/markdown-editor";
 
 import { MarkdownEditor } from "./markdown-editor";
