+++ apps/desktop/src/components/sidebar/sidebar-action-menu.tsx
