+++ apps/desktop/src/components/sidebar/sidebar.tsx
@@ -11,8 +11,8 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
-import { getDisplayFileName, normalizePath } from "@/lib/paths";
-import { getShortcutDisplay } from "@/shared/shortcuts";
+import { getDisplayFileName, normalizePath } from "@/core/paths";
+import { getShortcutDisplay } from "@/core/shortcuts";
 
 import type { NoteShortcutItem } from "@/types/navigation";
 import type {
@@ -21,16 +21,16 @@ import type {
   SidebarProps,
   SidebarRemoveTarget,
   SidebarSkillCollectionItem,
-} from "../types/sidebar";
+} from "../../types/sidebar";
 
-import { LogoComponent } from "./logo-component";
-import { SkillSourceLogo } from "./skill-source-logo";
+import { LogoComponent } from "../logo-component";
+import { SkillSourceLogo } from "../skills/skill-source-logo";
 import {
   SidebarActionMenu,
   buildFileMenuItems,
   type SidebarActionMenuCoords,
 } from "./sidebar-action-menu";
-import { ChevronRightIcon, MoreVerticalIcon, PinIcon, PlusIcon, FolderPlusIcon } from "./icons";
+import { ChevronRightIcon, MoreVerticalIcon, PinIcon, PlusIcon, FolderPlusIcon } from "../icons";
 import { SidebarTreeNode } from "./sidebar-tree-node";
 
 const normalizePathKey = (path: string) => normalizePath(path).toLowerCase();
