+++ apps/desktop/src/components/ui/input.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, type ComponentPropsWithoutRef } from "react";
 import { Input as InputPrimitive } from "@base-ui/react/input";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/core/utils";
 
 export type InputProps = ComponentPropsWithoutRef<"input">;
