+++ apps/desktop/src/components/ui/tooltip.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Tooltip as TooltipPrimitive } from "@base-ui/react/tooltip";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/core/utils";
 
 function TooltipProvider({ delay = 200, ...props }: TooltipPrimitive.Provider.Props) {
   return <TooltipPrimitive.Provider data-slot="tooltip-provider" delay={delay} {...props} />;
