+++ apps/desktop/src/components/update-notification.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 
-import type { AppInfo, UpdateState } from "@/shared/workspace";
+import type { AppInfo, UpdateState } from "@/core/workspace";
 
 import { Button } from "@/components/ui/button";
 import { Tooltip, TooltipTrigger, TooltipContent } from "@/components/ui/tooltip";
