+++ apps/desktop/src/core/errors.ts
