+++ apps/desktop/src/core/format-byte-size.ts
