+++ apps/desktop/src/core/note-navigation.ts
