+++ apps/desktop/src/core/paths.ts
