+++ apps/desktop/src/core/shortcuts.ts
