+++ apps/desktop/src/core/skill-agent-catalog.ts
