+++ apps/desktop/src/core/skills.ts
