+++ apps/desktop/src/core/workspace.ts
