+++ apps/desktop/src/hooks/use-sidebar-controller.ts
@@ -1,18 +1,18 @@
 import { useCallback, useEffect, useMemo, useState } from "react";
 
-import type { DirectoryNode, FileDocument, WorkspaceSnapshot } from "@/shared/workspace";
-import type { AppSettings } from "@/shared/workspace";
-import type { SidebarItemSetting } from "@/shared/workspace";
+import type { DirectoryNode, FileDocument, WorkspaceSnapshot } from "@/core/workspace";
+import type { AppSettings } from "@/core/workspace";
+import type { SidebarItemSetting } from "@/core/workspace";
 import type { DragPosition, SidebarTopLevelNode } from "@/types/sidebar";
 
-import { isPathInside, isSamePath, normalizePath } from "@/lib/paths";
+import { isPathInside, isSamePath, normalizePath } from "@/core/paths";
 import {
   filterSidebarNodes,
   removeSidebarPath,
   reorderSidebarNodes,
   toSidebarItemSetting,
   upsertSidebarFolder,
-} from "@/lib/sidebar-tree";
+} from "@/core/sidebar-tree";
 
 type UseSidebarControllerOptions = {
   glyph: NonNullable<Window["glyph"]>;
