+++ apps/desktop/src/store/workspace.ts
@@ -1,8 +1,8 @@
 import { create } from "zustand";
 
-import { getBaseName, isPathInside, isSamePath, normalizePath } from "@/lib/paths";
+import { getBaseName, isPathInside, isSamePath, normalizePath } from "@/core/paths";
 
-import type { DirectoryNode, FileDocument, NoteTab, TabMovePosition } from "../shared/workspace";
+import type { DirectoryNode, FileDocument, NoteTab, TabMovePosition } from "@/core/workspace";
 
 const getClosestHistoryIndex = (history: string[], currentIndex: number, filePath: string) => {
   let closestIndex = -1;
